fixed tiers regist, tier2 camera obs added

+++ skrl/skrl/agents/torch/dqn/dqn.py
@@ -194,7 +194,7 @@ class DQN(Agent):
 
         indexes = (torch.rand(states.shape[0], device=self.device) >= epsilon).nonzero().view(-1)
         if indexes.numel():
-            actions[indexes] = torch.argmax(self.q_network.act({"states": states[indexes]}, role="q_network")[0], dim=1, keepdim=True)
+            actions[indexes] = torch.argmax(self.q_network.act({"states": states[indexes]}, role="q_network")[0], dim=1, keepdim=True).float()  # Convert the source tensor to float dtype
 
         # record epsilon
         self.track_data("Exploration / Exploration epsilon", epsilon)
